Add monument faker

+++ database/migrations/2020_04_25_151857_create_monuments_table.php
@@ -17,12 +17,10 @@ class CreateMonumentsTable extends Migration
             $table->id();
 						$table->string('name');
 						$table->string('description');
-						$table->double('lat');
-						$table->double('lon');
+						$table->decimal('lat');
+						$table->decimal('lon');
 						$table->foreignId('user_id')->constrained()->onDelete('cascade');
 						$table->foreignId('image_id')->constrained()->onDelete('cascade');
-
-
             $table->timestamps();
         });
     }
